refactor: split pr lifecycle policy and runtime state
