+++ src/orchestrator/state.ts
@@ -1,17 +1,19 @@
 import type { RetryState } from "../domain/retry.js";
+import {
+  createFollowUpRuntimeState,
+  type FollowUpRuntimeState,
+} from "./follow-up-state.js";
 
 export interface OrchestratorState {
   readonly runningIssueNumbers: Set<number>;
   readonly retries: Map<number, RetryState>;
-  readonly nextAttemptByIssueNumber: Map<number, number>;
-  readonly followUpAttemptsByIssueNumber: Map<number, number>;
+  readonly followUp: FollowUpRuntimeState;
 }
 
 export function createOrchestratorState(): OrchestratorState {
   return {
     runningIssueNumbers: new Set<number>(),
     retries: new Map<number, RetryState>(),
-    nextAttemptByIssueNumber: new Map<number, number>(),
-    followUpAttemptsByIssueNumber: new Map<number, number>(),
+    followUp: createFollowUpRuntimeState(),
   };
 }
