+++ tests/unit/pull-request-policy.test.ts
@@ -1,0 +1,103 @@
+import { describe, expect, it } from "vitest";
+import { evaluatePullRequestLifecycle } from "../../src/tracker/pull-request-policy.js";
+import type { PullRequestSnapshot } from "../../src/tracker/pull-request-snapshot.js";
+
+function createSnapshot(
+  overrides?: Partial<PullRequestSnapshot>,
+): PullRequestSnapshot {
+  return {
+    branchName: "symphony/16",
+    pullRequest: {
+      number: 16,
+      url: "https://example.test/pulls/16",
+      branchName: "symphony/16",
+      latestCommitAt: "2026-03-06T00:00:00.000Z",
+    },
+    checks: [],
+    pendingCheckNames: [],
+    failingCheckNames: [],
+    actionableReviewFeedback: [],
+    botActionableReviewFeedback: [],
+    unresolvedThreadIds: [],
+    ...overrides,
+  };
+}
+
+describe("pull-request-policy", () => {
+  it("stabilizes no-check PRs before reporting ready", () => {
+    const snapshot = createSnapshot();
+
+    const first = evaluatePullRequestLifecycle(snapshot, undefined);
+    const second = evaluatePullRequestLifecycle(
+      snapshot,
+      first.nextNoCheckObservation ?? undefined,
+    );
+
+    expect(first.lifecycle.kind).toBe("awaiting-review");
+    expect(second.lifecycle.kind).toBe("ready");
+  });
+
+  it("waits on human-only review feedback without scheduling follow-up", () => {
+    const lifecycle = evaluatePullRequestLifecycle(
+      createSnapshot({
+        actionableReviewFeedback: [
+          {
+            id: "human-1",
+            kind: "review-thread",
+            threadId: "thread-1",
+            authorLogin: "jess",
+            body: "please clarify",
+            createdAt: "2026-03-06T00:00:00.000Z",
+            url: "https://example.test/thread/1",
+            path: "src/file.ts",
+            line: 10,
+          },
+        ],
+      }),
+      undefined,
+    ).lifecycle;
+
+    expect(lifecycle.kind).toBe("awaiting-review");
+    expect(lifecycle.actionableReviewFeedback).toHaveLength(1);
+  });
+
+  it("requires follow-up for failing checks or bot feedback", () => {
+    const lifecycle = evaluatePullRequestLifecycle(
+      createSnapshot({
+        failingCheckNames: ["CI"],
+        actionableReviewFeedback: [
+          {
+            id: "bot-1",
+            kind: "review-thread",
+            threadId: "thread-2",
+            authorLogin: "greptile[bot]",
+            body: "fix this",
+            createdAt: "2026-03-06T00:00:00.000Z",
+            url: "https://example.test/thread/2",
+            path: "src/file.ts",
+            line: 12,
+          },
+        ],
+        botActionableReviewFeedback: [
+          {
+            id: "bot-1",
+            kind: "review-thread",
+            threadId: "thread-2",
+            authorLogin: "greptile[bot]",
+            body: "fix this",
+            createdAt: "2026-03-06T00:00:00.000Z",
+            url: "https://example.test/thread/2",
+            path: "src/file.ts",
+            line: 12,
+          },
+        ],
+        unresolvedThreadIds: ["thread-2"],
+      }),
+      undefined,
+    ).lifecycle;
+
+    expect(lifecycle.kind).toBe("needs-follow-up");
+    expect(lifecycle.failingCheckNames).toEqual(["CI"]);
+    expect(lifecycle.unresolvedThreadIds).toEqual(["thread-2"]);
+  });
+});
